add challengeLink in Menu.js New Course method fixed refactor
