+++ modules/courses/_data/courseCreateData.js
@@ -1,0 +1,9 @@
+const courseRandom999 = Math.trunc(Math.random() * 1000);
+
+const courseData = {
+  name: courseRandom999 + ' Course for adding Challenge',
+  description: 'description, description',
+  type: 'all',
+};
+
+export { courseData };
