+++ modules/courses/courses_challenge/coursesChallengeNegativeNoSolution.spec.js
@@ -1,82 +1,69 @@
 import LoginPage from '../../user/_page/LoginPage';
 import LogoutPage from '../../user/_page/LogoutPage';
 import { admin, student } from '../../user/_data/user.data';
-import CoursesPage from '../../_PageObjects/CoursesPage';
-import ChallengePage from '../../_PageObjects/ChallengePage';
-import { courseData } from '../../_PageObjects/CoursesPage';
-import { challengeData } from '../../_PageObjects/ChallengePage';
-import Notification from '../../_PageObjects/Notification';
-import Menu from '../../_PageObjects/Menu';
+import CoursesPage from '../_page/CoursesPage';
+import ChallengePage from '../_page/ChallengePage';
+import { courseData } from '../_data/courseCreateData';
+import { challengeData } from '../_data/challengeCreateData';
+import Notification from '../../_page/Notification';
+import Menu from '../../_page/Menu';
 
-describe('ADD A NEW CHALLENGE TO CREATED COURSE ', () => {
+describe(' STUDENT COURSE CHALLENGE VALID INPUT', () => {
   before('should login as Admin, create Course, create Challenge', () => {
     LoginPage.login(admin);
     CoursesPage.createNewCourseGlobal;
     ChallengePage.createNewChallengeGlobal;
-    browser.pause(1000);
   });
 
-  it('should find and click on created course', () => {
-    CoursesPage.open('https://stage.pasv.us/course');
-    $(CoursesPage.listOfCourses).click();
-    browser.pause(1000);
+  it('should open on created course click EDIT', () => {
+    Menu.coursesLink.click();
+    CoursesPage.findCourse.click();
+    CoursesPage.editBtn.click();
   });
 
-  it('should click EDIT course', () => {
-    $(CoursesPage.editBtn).click();
-    browser.pause(1000);
-  });
-
-  it('should add created challenge to course', () => {});
+  //   it('should add created challenge to course', () => { });
 
   it('should logout', () => {
     LogoutPage.logout();
   });
-});
 
-describe('STUDENT CHECK CHALLENGE IN COURSE --- POSITIVE', () => {
-  before('should login as Student, open created course', () => {
+  it('should login as Student, open created course', () => {
     LoginPage.login(student);
-    CoursesPage.open('https://stage.pasv.us/course');
-    browser.pause(1000);
+    Menu.coursesLink.click();
+    CoursesPage.findCourse.click();
   });
 
-  it('should click on created course ', () => {
-    $(CoursesPage.findCourse).click();
-    browser.pause(1000);
-  });
-
-  it('should confirm h1', () => {
-    expect($(Menu.h1).getText()).eq(courseData.name);
-  });
-
-  it('should confirm describe', () => {
-    expect($(CoursesPage.courseDescription).getText()).eq(courseData.description);
+  it('should confirm h1, describe ', () => {
+    expect(Menu.h1.getText()).eq(courseData.name);
+    expect(CoursesPage.courseDescription.getText()).eq(courseData.description);
   });
 
   it('should click button "Start course" ', () => {
-    $(CoursesPage.startCourseBtn).click();
-    browser.pause(1000);
+    CoursesPage.startCourseBtn.click();
   });
 
-  it('should confirm challenge exist on the page and has same name as created', () => {
-    expect(ChallengePage.challengeInsideCourseName.getText()).eq(challengeData.name);
-  });
-
-  it('should fill in correct solution in "Write your solution here" textarea', () => {
-    $(ChallengePage.challengeInsideCourseTextarea).setValue('');
-  });
-
-  it('should click "Validate solution" ', () => {
-    $(ChallengePage.challengeInsideCourseValidateBtn).click();
-    browser.pause(1000);
-  });
-
-  it('should confirm tests NOT passed', () => {
-    expect($(ChallengePage.challengePassArray).length).notEqual(challengeData.testsQuantity);
-  });
-
-  it('should confirm fail notification wrapper is Displayed', () => {
-    $(Notification.fail).isDisplayed();
-  });
+  // it('should confirm challenge name', () => {
+  //   expect(ChallengePage.challengeInsideCourseName.getText()).eq(challengeData.name);
+  // });
+  //
+  // it('should fill in WRONG solution in "Write your solution here" textarea', () => {
+  //   ChallengePage.challengeInsideCourseTextarea.setValue('');
+  // });
+  //
+  // it('should click button "Validate solution" ', () => {
+  //   ChallengePage.challengeInsideCourseValidateBtn.click();
+  //   browser.pause(1000);
+  // });
+  //
+  // it('should confirm tests NOT passed', () => {
+  //   expect(ChallengePage.challengePassArray.length).notEqual(challengeData.testsQuantity);
+  // });
+  //
+  // it('should confirm FAIL notification wrapper is Displayed', () => {
+  //   expect(Notification.fail.isDisplayed()).true;
+  // });
+  //
+  // it('should confirm challenge NOT marked as a "Completed"', () => {
+  //   ChallengePage.challengeCompleteMark.isDisplayed().false;
+  // });
 });
