+++ modules/courses/courses_progress/coursesProgress.spec.js
@@ -5,7 +5,7 @@ import ProfilePage from '../../user/_page/ProfilePage';
 import CoursesProgressPage from '../../_page/CoursesProgressPage';
 import { data } from '../../user/_data/profilePage.data';
 import { admin, student } from '../../user/_data/user.data';
-import CoursesPage from '../../_page/CoursesPage';
+import CoursesPage from '../_page/CoursesPage';
 import { course } from '../../_data/courseProgress.data';
 
 // this test will be refactored once Create New Course method is proper
