+++ package.json
@@ -33,5 +33,8 @@
   },
   "dependencies": {
     "chai": "^4.2.0"
+  },
+  "directories": {
+    "doc": "docs"
   }
 }
